backend/requirements.txt
# ============================================================
# Dependencias — Proyectos CJDG Backend (Python / FastAPI)
# ============================================================

# Framework principal
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Base de datos
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# Validación de datos (Pydantic v2)
pydantic==2.7.1
pydantic-settings==2.2.1

# Autenticación JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9

# Variables de entorno
python-dotenv==1.0.1

# Procesamiento de PDFs y archivos estáticos
pdfplumber==0.11.4
Pillow==10.3.0
